Consolidate DashboardView contexts
JIRA: RAIL-2956
