+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/InsightRenderer.tsx
@@ -32,10 +32,12 @@ import { InsightError } from "../../../insightView/InsightError";
 import { widgetDrillsToDrillPredicates, insightDrillDownPredicates } from "./drillingUtils";
 import { filterContextItemsToFiltersForWidget, filterContextToFiltersForWidget } from "../converters";
 import { addImplicitAllTimeFilter } from "./utils";
-import { useDashboardViewConfig } from "./DashboardViewConfigContext";
-import { useUserWorkspaceSettings } from "./UserWorkspaceSettingsContext";
-import { useColorPalette } from "./ColorPaletteContext";
-import { useAttributesWithDrillDown } from "./AttributesWithDrillDownContext";
+import {
+    useAttributesWithDrillDown,
+    useColorPalette,
+    useDashboardViewConfig,
+    useUserWorkspaceSettings,
+} from "./contexts";
 
 interface IInsightRendererProps {
     insightWidget: IInsightWidget;
