+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/KpiView/alertManipulationHooks/useAlertSaveOrUpdateHandler.ts
@@ -3,7 +3,7 @@ import { useState } from "react";
 import { IAnalyticalBackend, isWidgetAlert, IWidgetAlertDefinition } from "@gooddata/sdk-backend-spi";
 import { useSaveOrUpdateWidgetAlert } from "../../../hooks/useSaveOrUpdateWidgetAlert";
 import { KpiAlertOperationStatus } from "../../../types";
-import { useAlerts } from "../../DashboardAlertsContext";
+import { useAlerts } from "../../contexts";
 
 interface IUseAlertSaveHandlerConfig {
     closeAlertDialog: () => void;
