+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/DashboardView/contexts/DashboardViewIsReadOnlyContext.tsx
